pyproject.toml
[tool.poetry]
name = "glyph"
version = "0.1.0"
description = "A highly customizable shell prompt"
authors = ["Ian Baldwin <ian@iantbaldw.in>"]

[tool.poetry.dependencies]
python = "^3.8"
sh = "^1.13.1"
logzero = "^1.5.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
bandit = "^1.6.2"
mypy = "^0.782"
pylint = "^2.5.3"
black = "^19.10b0"

[tool.poetry.scripts]
glyph = 'glyph.cli:cli'

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
